Transcendence/src/backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.3
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
wheel==0.42.0
requests==2.31.0
pillow==10.2.0
django-extensions==3.2.3
Werkzeug==3.0.1
pyOpenSSL==24.0.0
gunicorn==21.2.0
pyotp==2.9.0
qrcode==7.4.2
channels==4.0.0
daphne==4.1.0
